refactor overview and agent dashboard
