+++ client/src/Demo/TaiGerOrg/InternalDashboard/index.js
@@ -1,43 +1,16 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import {
-  Card,
-  Tabs,
-  Tab,
-  Box,
-  Typography,
-  Grid,
-  Breadcrumbs,
-  ButtonGroup,
-  Button
-} from '@mui/material';
+import { Tabs, Tab, Box, Typography, Breadcrumbs } from '@mui/material';
 import PropTypes from 'prop-types';
 import { Navigate, Link as LinkDom, useLocation } from 'react-router-dom';
 import { Link } from '@mui/material';
-import { LineChart } from '@mui/x-charts/LineChart';
-
-import {
-  BarChart,
-  CartesianGrid,
-  Tooltip,
-  XAxis,
-  YAxis,
-  Bar,
-  Cell,
-  ResponsiveContainer
-} from 'recharts';
 import { useTranslation } from 'react-i18next';
-import { BarChart as MuiBarChart } from '@mui/x-charts/BarChart';
 import dayjs from 'dayjs';
 import isoWeek from 'dayjs/plugin/isoWeek';
 import { is_TaiGer_role } from '@taiger-common/core';
 dayjs.extend(isoWeek);
 
 import ErrorPage from '../../Utils/ErrorPage';
-import {
-  isProgramDecided,
-  isProgramSubmitted,
-  open_tasks_with_editors
-} from '../../Utils/checking-functions';
+
 import { getStatistics } from '../../../api';
 import { TabTitle } from '../../Utils/TabTitle';
 import DEMO from '../../../store/constant';
@@ -50,99 +23,11 @@ import {
   INTERNAL_DASHBOARD_REVERSED_TABS,
   INTERNAL_DASHBOARD_TABS
 } from '../../Utils/contants.js';
-import SingleBarChart from '../../../components/Charts/SingleBarChart';
-import VerticalDistributionBarCharts from '../../../components/Charts/VerticalDistributionBarChart';
-import VerticalSingleBarChart from '../../../components/Charts/VerticalSingleChart';
 
+import OverviewDashboard from './OverviewDashboard.js';
+import AgentDashboard from './AgentDashboard';
 import KPIDashboard from './KPIDashboard';
 import ResponseTimeDashboard from './ResponseTimeDashboard';
-
-function groupByMonth(data) {
-  return data.reduce((acc, { createdAt }) => {
-    const month = dayjs(createdAt).format('YYYY-MM');
-    if (!acc[month]) {
-      acc[month] = 0;
-    }
-    acc[month]++;
-    return acc;
-  }, {});
-}
-
-function groupByWeek(data) {
-  return data.reduce((acc, { createdAt }) => {
-    const week = dayjs(createdAt).isoWeek();
-    const year = dayjs(createdAt).year();
-    const weekYear = `${year}-W${week}`;
-    if (!acc[weekYear]) {
-      acc[weekYear] = 0;
-    }
-    acc[weekYear]++;
-    return acc;
-  }, {});
-}
-
-function prepareChartData(groupedData) {
-  const labels = Object.keys(groupedData).sort((a, b) => {
-    if (a.includes('-W') && b.includes('-W')) {
-      const [yearA, weekA] = a.split('-W').map(Number);
-      const [yearB, weekB] = b.split('-W').map(Number);
-      return yearA === yearB ? weekA - weekB : yearA - yearB;
-    }
-    return a.localeCompare(b);
-  });
-
-  const data = labels.map((label) => groupedData[label]);
-
-  return {
-    labels,
-    datasets: [
-      {
-        label: 'Data Points',
-        data
-      }
-    ]
-  };
-}
-
-const AgentBarCharts = ({ agentDistr }) => {
-  // Extract unique years from both arrays
-
-  const combinedKeys = Array.from(
-    new Set(
-      [
-        ...Object.keys(agentDistr.admission),
-        ...Object.keys(agentDistr.noAdmission)
-      ]?.sort()
-    )
-  );
-  const admissionDataset = [];
-  const noAdmissionDataset = [];
-
-  // Populate datasets
-  combinedKeys.forEach((key) => {
-    admissionDataset.push(agentDistr.admission[key] || 0);
-    noAdmissionDataset.push(agentDistr.noAdmission[key] || 0);
-  });
-  return (
-    <Box>
-      <Typography variant="h6">{agentDistr.name}</Typography>
-      <MuiBarChart
-        xAxis={[{ scaleType: 'band', data: combinedKeys }]}
-        yAxis={[
-          {
-            label: 'Student'
-          }
-        ]}
-        series={[
-          { label: 'No Offer', data: noAdmissionDataset },
-          { label: 'Has Offer', data: admissionDataset }
-        ]}
-        height={250}
-        width={400}
-      />
-    </Box>
-  );
-};
 
 CustomTabPanel.propTypes = {
   children: PropTypes.node,
@@ -238,7 +123,7 @@ function InternalDashboard() {
   const { user } = useAuth();
   const { t } = useTranslation();
   const { hash } = useLocation();
-  const [viewMode, setViewMode] = useState('month');
+
   const memoizedColumnsMrt = useMemo(
     () => responseTimeColumn,
     [responseTimeColumn]
@@ -348,23 +233,6 @@ function InternalDashboard() {
   if (res_status >= 400) {
     return <ErrorPage res_status={res_status} />;
   }
-  const colors = [
-    '#ff8a65',
-    '#f4c22b',
-    '#04a9f5',
-    '#3ebfea',
-    '#4F5467',
-    '#1de9b6',
-    '#a389d4',
-    '#FE8A7D'
-  ];
-
-  const groupedData =
-    viewMode === 'month'
-      ? groupByMonth(students_details)
-      : groupByWeek(students_details);
-
-  const chartData = prepareChartData(groupedData);
 
   // Normalize data by ensuring each object has all keys
   const normalizedResults = studentResponseTimeLookupTable.map((result) => {
@@ -373,93 +241,6 @@ function InternalDashboard() {
       normalizedResult[key] = result[key] !== undefined ? result[key] : 0; // Set missing keys to null or undefined
     });
     return normalizedResult;
-  });
-  console.log(normalizedResults);
-  const application_status = ['Open', 'Close'];
-  const admission_status = ['Admission', 'Rejection', 'Pending'];
-
-  const initialData = {
-    Open: 0,
-    Close: 0,
-    Admission: 0,
-    Rejection: 0,
-    Pending: 0
-  };
-
-  const data = students_details.reduce((acc, student) => {
-    student.applications.forEach((application) => {
-      if (application.program_id !== '-' && isProgramDecided(application)) {
-        if (isProgramSubmitted(application)) {
-          acc.Close += 1;
-          if (application.admission === 'O') {
-            acc.Admission += 1;
-          } else if (application.admission === 'X') {
-            acc.Rejection += 1;
-          } else if (application.admission === '-') {
-            acc.Pending += 1;
-          }
-        } else {
-          acc.Open += 1;
-        }
-      }
-    });
-    return acc;
-  }, initialData);
-
-  const applications_data = application_status.map((status) => ({
-    name: status,
-    uv: data[status]
-  }));
-
-  const admissions_data = admission_status.map((status) => ({
-    name: status,
-    uv: data[status]
-  }));
-
-  // Only open tasks. Closed tasks are excluded
-  const open_tasks_arr = open_tasks_with_editors(students_details);
-
-  const documents_data = [];
-  const editor_tasks_distribution_data = [];
-  const cat = [
-    'CURRICULUM_ANALYSIS',
-    'CV',
-    'ESSAY',
-    'FORM_A',
-    'FORM_B',
-    'INTERNSHIP_FORM',
-    'ML',
-    'OTHERS',
-    'PORTFOLIO',
-    'RL',
-    'SCHOLARSHIP_FORM',
-    'SUPPLEMENTARY_FORM'
-  ];
-  cat.forEach((ca) => {
-    documents_data.push({
-      name: `${ca}`,
-      uv: documents[ca]?.count || 0
-      // color: colors[i]
-    });
-  });
-  editors_data.forEach((editor) => {
-    editor_tasks_distribution_data.push({
-      name: `${editor.firstname}`,
-      active: open_tasks_arr.filter(
-        ({ editors, isFinalVersion, show }) =>
-          editors.findIndex((ed) => ed._id === editor._id) !== -1 &&
-          isFinalVersion !== true &&
-          show
-      ).length,
-      potentials: open_tasks_arr.filter(
-        ({ editors, isFinalVersion, show, isPotentials }) =>
-          editors.findIndex((ed) => ed._id == editor._id) !== -1 &&
-          isFinalVersion !== true &&
-          !show &&
-          isPotentials
-      ).length
-      // color: colors[i]
-    });
   });
 
   const calculateDuration = (start, end) => {
@@ -583,179 +364,20 @@ function InternalDashboard() {
         </Tabs>
       </Box>
       <CustomTabPanel value={value} index={0}>
-        <Grid container spacing={2}>
-          <Grid item xs={12}>
-            <Card sx={{ p: 2 }}>
-              <ButtonGroup
-                variant="contained"
-                aria-label="outlined primary button group"
-                style={{ marginBottom: '20px' }}
-              >
-                <Button
-                  onClick={() => setViewMode('month')}
-                  variant={viewMode === 'month' ? 'contained' : 'outlined'}
-                >
-                  Month View
-                </Button>
-                <Button
-                  onClick={() => setViewMode('week')}
-                  variant={viewMode === 'week' ? 'contained' : 'outlined'}
-                >
-                  Week View
-                </Button>
-              </ButtonGroup>
-              <LineChart
-                xAxis={[{ data: chartData.labels, scaleType: 'band' }]}
-                series={[
-                  {
-                    data: chartData.datasets[0].data
-                  }
-                ]}
-                yAxis={[
-                  {
-                    label: 'New Students'
-                  }
-                ]}
-                height={300}
-              />
-            </Card>
-          </Grid>
-          <Grid item xs={12} md={4}>
-            <Card>
-              <Typography>{t('Students')}</Typography>
-              <SingleBarChart
-                data={students_years_pair}
-                label="Student"
-                yLabel="Student"
-              />
-            </Card>
-          </Grid>
-          <Grid item xs={12} md={4}>
-            <Card sx={{ p: 2 }}>
-              <Typography>Tasks: Number of Tasks</Typography>
-              <SingleBarChart
-                data={documents_data}
-                label="Tasks"
-                yLabel="Tasks"
-              />
-            </Card>
-          </Grid>
-          <Grid item xs={12} md={4}>
-            <Card sx={{ p: 2 }}>
-              <Typography>
-                {t('Agents', { ns: 'common' })}: Number of active students per
-                agent
-              </Typography>
-              <VerticalDistributionBarCharts
-                data={agents_data}
-                k={'key'}
-                value1={'student_num_no_offer'}
-                value2={'student_num_with_offer'}
-                xLabel="Student"
-                dataALabel="No Offer"
-                dataBLabel="Has Offer"
-              />
-            </Card>
-          </Grid>
-        </Grid>
-        <Grid container spacing={2}>
-          <Grid item xs={12} md={4}>
-            <Card sx={{ p: 2 }}>
-              <Typography>
-                {t('Editor', { ns: 'common' })}: Number of active students per
-                editor
-              </Typography>
-              <VerticalSingleBarChart data={editors_data} xLabel={'Student'} />
-            </Card>
-          </Grid>
-          <Grid item xs={12} md={4}>
-            <Card sx={{ p: 2 }}>
-              <Typography>
-                {t('Editor', { ns: 'common' })}:Number of open and potential
-                tasks per editor
-              </Typography>
-              <VerticalDistributionBarCharts
-                data={editor_tasks_distribution_data}
-                k={'name'}
-                value1={'active'}
-                value2={'potentials'}
-                xLabel="Tasks"
-                dataALabel="Active"
-                dataBLabel="Potentials"
-              />
-            </Card>
-          </Grid>
-          <Grid item xs={12} md={4}>
-            <Card sx={{ p: 2 }}>
-              <Typography>
-                {t('Applications')}: Number of Applications:
-              </Typography>
-              <ResponsiveContainer width="100%" height={300}>
-                <BarChart
-                  data={applications_data}
-                  margin={{
-                    top: 20,
-                    right: 0,
-                    left: 0,
-                    bottom: 5
-                  }}
-                >
-                  <CartesianGrid strokeDasharray="3 3" />
-                  <XAxis dataKey="name" />
-                  <YAxis allowDecimals={false} />
-                  <Tooltip labelStyle={{ color: 'black' }} />
-                  {/* <Legend /> */}
-                  <Bar dataKey="uv" fill="#8884d8" label={{ position: 'top' }}>
-                    {applications_data.map((entry, index) => (
-                      <Cell key={`cell-${index}`} fill={colors[index % 20]} />
-                    ))}
-                  </Bar>
-                </BarChart>
-              </ResponsiveContainer>
-            </Card>
-          </Grid>
-          <Grid item xs={12} md={4}>
-            <Card sx={{ p: 2 }}>
-              <Typography>{t('Admissions')}: Number of Admissions</Typography>
-              <ResponsiveContainer width="100%" height={300}>
-                <BarChart
-                  data={admissions_data}
-                  margin={{
-                    top: 20,
-                    right: 0,
-                    left: 0,
-                    bottom: 5
-                  }}
-                >
-                  <CartesianGrid strokeDasharray="3 3" />
-                  <XAxis dataKey="name" />
-                  <YAxis allowDecimals={false} />
-                  <Tooltip labelStyle={{ color: 'black' }} />
-                  {/* <Legend /> */}
-                  <Bar dataKey="uv" fill="#8884d8" label={{ position: 'top' }}>
-                    {admissions_data.map((entry, index) => (
-                      <Cell key={`cell-${index}`} fill={colors[index % 20]} />
-                    ))}
-                  </Bar>
-                </BarChart>
-              </ResponsiveContainer>
-            </Card>
-          </Grid>
-        </Grid>
+        <OverviewDashboard
+          studentDetails={students_details}
+          agentData={agents_data}
+          editorData={editors_data}
+          studentsYearsPair={students_years_pair}
+          documents={documents}
+        />
       </CustomTabPanel>
       <CustomTabPanel value={value} index={1}>
-        <Grid container spacing={2}>
-          <Grid item xs={12}>
-            <Typography variant="h6">Active Students distribution</Typography>
-          </Grid>
-          {internalDashboardState.agentStudentDistribution.map(
-            (agentDistr, idx) => (
-              <Grid item xs={12} md={4} key={idx}>
-                <AgentBarCharts agentDistr={agentDistr} />
-              </Grid>
-            )
-          )}
-        </Grid>
+        <AgentDashboard
+          agentStudentDistribution={
+            internalDashboardState.agentStudentDistribution
+          }
+        />
       </CustomTabPanel>
       <CustomTabPanel value={value} index={2}>
         <KPIDashboard
